# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medilink Healthcare API",
    "description": "Comprehensive Interactive API Documentation for Medilink Healthcare Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "/",
      "description": "Relative server route"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT authentication token."
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/emergency/sos-update": {
      "patch": {
        "summary": "Update SOS Alert Status",
        "tags": ["Emergency"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["request_id", "status"],
                "properties": {
                  "request_id": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" },
                  "status": { "type": "string", "example": "DISPATCHED" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
    "/api/emergency/create": {
      "post": {
        "summary": "Create Emergency Case",
        "tags": ["Emergency"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["patient_name", "phone"],
                "properties": {
                  "patient_name": { "type": "string", "example": "Kiran Rahul" },
                  "phone": { "type": "string", "example": "9123456780" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/emergency/update-status": {
      "patch": {
        "summary": "Update Emergency Case Status",
        "tags": ["Emergency"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["case_id", "status"],
                "properties": {
                  "case_id": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" },
                  "status": { "type": "string", "example": "RESOLVED" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/emergency/assign-bed": {
      "patch": {
        "summary": "Assign Bed to Case",
        "tags": ["Emergency"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["case_id", "bed_id"],
                "properties": {
                  "case_id": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" },
                  "bed_id": { "type": "string", "format": "uuid", "example": "bb0b4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
  }
}